fix(ui): ensure event handlers return new page val
all callbacks now return the new page value if it exists. If not then
the event is returned.

Clean up inputchange, keypress, blur handlers. There was some weirdness
going on. They all messed with the internal state. Changed this so only
the inputChange handler sets a new state. The others just take
whatever's in the state and use it when firing the callbacks.

Additionally there was an issue that the internal state wasn't always
updated correctly. The three handlers had slightly different logic and
could really only function in conjunction. E.g. the keyPress handler
only dealt with the edge cases (value too low or high) but not with the
happy path, whereas the inputHandler just set the new state
indiscriminantly even if it was too high or too low.