feat(ui): fixes Pagination inconsistencies. Adds disabled and loading state (#352)
* Update packages/ui-components/src/components/Pagination/Pagination.component.js

Co-authored-by: Esther Schmitz <edda@users.noreply.github.com>

* fix(ui): fixes behavior with undefined 'currentPages' or/and 'totalPages'

* fix(ui): fixes controlPage wasn't increased if controlTotalPage is undefined

* fix(ui): fixes handlePressNext

* fix(ui): fixes onPressNext fires even if controlPage is undefined

* fix(ui): ensure event handlers return new page val

all callbacks now return the new page value if it exists. If not then
the event is returned.

Clean up inputchange, keypress, blur handlers. There was some weirdness
going on. They all messed with the internal state. Changed this so only
the inputChange handler sets a new state. The others just take
whatever's in the state and use it when firing the callbacks.

Additionally there was an issue that the internal state wasn't always
updated correctly. The three handlers had slightly different logic and
could really only function in conjunction. E.g. the keyPress handler
only dealt with the edge cases (value too low or high) but not with the
happy path, whereas the inputHandler just set the new state
indiscriminantly even if it was too high or too low.

* fix(ui): fixes next/previous Click only change state if condition is true, Input change fires event

* fix(ui): correct selectChange and inputChange handler comments

* refactor(ui): events always return currentPage or undefined, never event

* fix(ui): remove unused vars

* feat(ui): add changeset

---------

Co-authored-by: Wowa <vladimir.barsukov@sap.com>
Co-authored-by: Andreas Pfau <andreas.pfau@sap.com>
Co-authored-by: D038721 <esther.schmitz@sap.com>